requirements.txt
streamlit
youtube-transcript-api
yt-dlp
openai-whisper
g4f
pandas